kkkkkkkkkkkkkkkkkkkkkkkk

+++ scripts/stop_server.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
-echo ahahahh
-kill $(pgrep uvicorn)
 
+uvicorn_pid=`pgrep uvicorn`
+if [ ! -z "$uvicorn_pid" ]; then 
+	kill $(pgrep uvicorn)
+fi
